VIX-2624 Incremental commit with more progress on the element filter editor.
More work on the models and views for the filter editor. Refactored the HasSetupVisibilityConverter from the Layer editor and made it more generic to work with any object that supports the IHasSetup interface. This allows it to work with layers and these filters. There are a few more items that can be refactored in the layer editor at some point. Fixed the effect instance to use the new wrapper classes.