VIX-2624 Bug fixes in the filter / transform editor.
Made some fixes so that when changes happen to the filter setup, proper rendering update events occur. Changed a few visual notions from filter to transform to see how that feels from a user perspective. If that name is adopted, the underlying code naming should be updated. Add logic to hide the filter list property to be visible in the effect editor.