VIX-2624 Add some additional logic into the chase effect for target node handling.
Certain effects will need to provide additional functionality to manage how they see the new transformed target nodes. In the case of the Chase effect, they may want a chase applied to each individual transformed element, or they may want the to be viewed as a group. This introduces that notion and allows even more flexibility in configuring how the effect behaves.