VIX-2624 Baseline work to allow for virtual element nodes to exist.
Added interfaces for IElementNode and IGroupNode to allow for multiple types of each. The current ElementNode and GroupNode objects interact with the system registry and we will need to have versions of them that just exist without altering the core setup. Refactored all the areas in the effect path that utilize the ElementNode to interact with the more generic interfaces so they do not care which version they get. This is a significant refactoring and as such will need to healthy testing.

Created the new VirtualElementNode class that we should be able to use when we need virtual element grouping. Will likely need to create a manager class similar to the NodeManager to help enforce structure. We will see if there is a need for that.