VIX-2705 Fix up a bunch of stuff after the latest rebase to get the world back working again.
Update new projects to 4.7.2 so they will play with the rest of them.

Fix some missing references to the Video property that somehow did not make the rebase.

Fix the smart controller policy to work with the latest changes in how state is moved. This needs some major rework yet as the concept of what it is doing here is no longer valid.