VIX-759 Finish the refactor to use CSCore Sound library.
Remove NAudio.
Refactor the sample providers to utilize CSCore.
Add ability to save the preferred audio Device.
Create a IPlayer interface to mange the interactions so future changes might be easier.
Remove FMOD libraries and wrapper files.
Remove the old Mark Manager that was heavily reliant on the FMOD details.